Found a replacement for CasEventListener
