+++ src/main/java/org/jenkinsci/plugins/cas/CasSecurityRealm.java
@@ -29,14 +29,16 @@ import hudson.Util;
 import hudson.model.Descriptor;
 import hudson.security.ChainedServletFilter;
 import hudson.security.SecurityRealm;
+import hudson.tasks.Mailer;
 import hudson.util.FormValidation;
+import java.util.Collection;
+import java.util.Map;
 import jenkins.model.Jenkins;
 import jenkins.security.SecurityListener;
 import org.jasig.cas.client.session.HashMapBackedSessionMappingStorage;
 import org.jasig.cas.client.session.SingleSignOutHandler;
 import org.jasig.cas.client.validation.AbstractUrlBasedTicketValidator;
 import org.jasig.cas.client.validation.TicketValidator;
-import org.jenkinsci.plugins.cas.spring.CasEventListener;
 import org.jenkinsci.plugins.cas.spring.security.CasAuthenticationEntryPoint;
 import org.jenkinsci.plugins.cas.spring.security.CasSingleSignOutFilter;
 import org.jenkinsci.plugins.cas.spring.security.CasUserDetailsService;
@@ -156,12 +158,6 @@ public class CasSecurityRealm extends SecurityRealm {
 	 */
 	protected void init() {
 		if (casAuthenticationEntryPoint == null) {
-            CasEventListener casEventListener;
-            { // TODO How can this be used? With the DefaultAuthenticationEventPublisher below?
-                casEventListener = new CasEventListener();
-                casEventListener.setFullNameAttribute(casProtocol.getFullNameAttribute());
-                casEventListener.setEmailAttribute(casProtocol.getEmailAttribute());
-            }
             ServiceProperties casServiceProperties;
             {
                 casServiceProperties = casProtocol.createServiceProperties();
@@ -371,14 +367,73 @@ public class CasSecurityRealm extends SecurityRealm {
 	 * @param req request
 	 * @param rsp response
 	 */
-	public void doFinishLogin(StaplerRequest req, StaplerResponse rsp) {
+	public void doFinishLogin(StaplerRequest req, StaplerResponse rsp) throws IOException {
 		org.springframework.security.core.Authentication authentication = org.springframework.security.core.context.SecurityContextHolder.getContext().getAuthentication();
 		if (authentication instanceof CasAuthenticationToken) {
 			org.springframework.security.core.context.SecurityContextHolder.clearContext();
 			CasAuthenticationToken casToken = (CasAuthenticationToken) authentication;
 			SecurityContextHolder.getContext().setAuthentication(casToken);
+			syncUserAttributes(casToken);
 			SecurityListener.fireAuthenticated2(casToken.getUserDetails());
 		}
+	}
+
+	/**
+	 * Sync user attributes with a CAS authentication token.
+	 * @param casToken CAS authentication token
+	 * @throws IOException
+	 */
+	protected void syncUserAttributes(CasAuthenticationToken casToken) throws IOException {
+		if (casToken.getAssertion() == null ||
+			casToken.getAssertion().getPrincipal() == null ||
+			casToken.getAssertion().getPrincipal().getAttributes() == null) {
+			// No attributes to sync with
+			return;
+		}
+
+		// Retrieve or create the Jenkins user
+		hudson.model.User user = hudson.model.User.get(casToken.getName());
+
+		// Sync the full name
+		String fullName = getAttributeValue(casToken, casProtocol.getFullNameAttribute());
+		if (fullName != null) {
+			user.setFullName(fullName);
+		}
+
+		// Sync the email address
+		String email = getAttributeValue(casToken, casProtocol.getEmailAttribute());
+		if (email != null) {
+			user.addProperty(new Mailer.UserProperty(email));
+		}
+
+		// Save the user to disk
+		user.save();
+	}
+
+	/**
+	 * Retrieve an attribute's value from a CAS authentication token.
+	 * @param authToken CAS authentication token
+	 * @param attributeName attribute name
+	 * @return attribute value or null if not found
+	 */
+	@SuppressWarnings("rawtypes")
+	protected String getAttributeValue(CasAuthenticationToken authToken, String attributeName) {
+		if (authToken != null && authToken.getAssertion() != null &&
+			authToken.getAssertion().getPrincipal() != null &&
+			authToken.getAssertion().getPrincipal().getAttributes() != null &&
+			attributeName != null) {
+
+			Map attributes = authToken.getAssertion().getPrincipal().getAttributes();
+			Object attribute = attributes.get(attributeName);
+			if (attribute != null) {
+				if (attribute instanceof Collection) {
+					return ((Collection) attribute).iterator().next().toString();
+				} else {
+					return attribute.toString();
+				}
+			}
+		}
+		return null;
 	}
 
 	// ~ SecurityRealm descriptor =======================================================================================
